Modified Tester and Trainer to work with np arrays
These programs assume that the input data set (data coordinates and labels) will come in the form of np arrays. Trainer batches the data according to the user entered batch size.

Also added placeholders for the weights and biases of HL1 and the Output Layer. This creates a place for the weights and biases to be inserted into the comp graphs when they are run.