GeneratingSquare.py
This generates a square with Gaussian distributions perpendicular to its edges. The user can enter the side length, the sigma of the Gaussian, and the total number of points desired.

To-do: Find a way to calculate step size as a function of total number of points and number of sides s.t. the points are distributed appropriately across the edge.