Works with new data generation.
The new generation is different from the old one in that it allows the user to input the area, number of distributions per side, and number of points per distribution. The total number of points is now calculated instead of being a command line argument to ensure that all sides of the square have an equal number of points.